Fixed with restrictions on playing music in the voice (random selection)

+++ commands/speaker_cog/music.py
@@ -5,9 +5,9 @@ import subprocess
 import disnake
 import logging
 from disnake.ext import commands, tasks
-from BANNED_FILES.config import SPEAKER_VOICE_ID, Music_Folder, Volume_Music
+from BANNED_FILES.config import Music_Folder, Volume_Music, RedisManager
+from redis_storage.speaker_voice import SpeakerVoice
 
-# Отключаем спам от Disnake в логах по voice_client
 logging.getLogger("disnake.voice_client").setLevel(logging.CRITICAL)
 
 class MusicPlayer(commands.Cog):
@@ -39,17 +39,34 @@ class MusicPlayer(commands.Cog):
                 self.voice_client = None
                 self.last_disconnect_time = asyncio.get_event_loop().time()
 
+    async def get_voice_channel(self) -> disnake.VoiceChannel | None:
+        """Получаем текущую голосовую руму из Redis по ключу"""
+        async with RedisManager() as redis:
+            try:
+                record = await redis.load(SpeakerVoice, key="random_channel")
+            except Exception as e:
+                print(f"[MusicPlayer] Ошибка при загрузке из Redis: {e}")
+                return None
+
+        if not record or not record.random_channel_id:
+            print("[MusicPlayer] Голосовая рума не найдена в Redis")
+            return None
+        channel = self.bot.get_channel(int(record.random_channel_id))
+        if not isinstance(channel, disnake.VoiceChannel):
+            print("[MusicPlayer] Канал в Redis не является голосовым")
+            return None
+        return channel
+
     async def connect_and_play(self):
         async with self.connecting_lock:
             if self.integration_cog is None:
                 self.integration_cog = self.bot.get_cog("MusicIntegration")
 
-            voice_channel = self.bot.get_channel(SPEAKER_VOICE_ID)
-            if not isinstance(voice_channel, disnake.VoiceChannel):
+            voice_channel = await self.get_voice_channel()
+            if not voice_channel:
                 print("[MusicPlayer] Канал не найден или не является голосовым")
                 return
 
-            # Если недавно отключались — подождать минимум 30 секунд
             if self.last_disconnect_time is not None:
                 elapsed = asyncio.get_event_loop().time() - self.last_disconnect_time
                 if elapsed < 30:
@@ -60,18 +77,17 @@ class MusicPlayer(commands.Cog):
                     if self.voice_client.channel.id != voice_channel.id:
                         await self.voice_client.move_to(voice_channel)
                 else:
-                    # Принудительно отключаем, если надо
                     await self.force_disconnect()
                     self.voice_client = await voice_channel.connect()
 
                 if self.integration_cog:
-                    await self.integration_cog.send_or_update_message("Ожидание пожалуста музыки....")
+                    await self.integration_cog.send_or_update_message("Ожидание музыки...")
 
             except disnake.ClientException as e:
                 print(f"[MusicPlayer] Ошибка подключения к голосовому каналу: {e}")
                 return
 
-            await asyncio.sleep(15)  # Можно подстроить задержку, если нужно
+            await asyncio.sleep(5)
 
             files = [f for f in os.listdir(self.music_folder) if f.lower().endswith((".mp3", ".wav", ".ogg", ".aac"))]
             if not files:
@@ -112,12 +128,11 @@ class MusicPlayer(commands.Cog):
 
     @tasks.loop(seconds=30)
     async def auto_reconnect(self):
-        voice_channel = self.bot.get_channel(SPEAKER_VOICE_ID)
-        if isinstance(voice_channel, disnake.VoiceChannel):
-            if self.is_disconnected():
-                if self.integration_cog:
-                    await self.integration_cog.delete_message()
-                await self.connect_and_play()
+        voice_channel = await self.get_voice_channel()
+        if voice_channel and self.is_disconnected():
+            if self.integration_cog:
+                await self.integration_cog.delete_message()
+            await self.connect_and_play()
 
     @auto_reconnect.before_loop
     async def before_loop(self):
